requirements.txt
torch
gradio
transformers
Pillow
opencv-python
numpy
timm